feat(landing): add landing page, move dashboard to /dashboard
The dashboard occupied /, so the project had no public front door and no way
to explain itself to either audience before asking them to read a table.

Routes now:
  /           landing page (static, server-rendered)
  /dashboard  merchant view, unchanged
  /verify     public receipt verifier

The landing page leads with the problem rather than a feature list — the
agent cannot audit a receipt issued by the seller's own API, and the merchant
cannot refund without holding an unbounded key over their own float. It then
shows the three-step flow, links both deployed contracts to Stellar Expert,
splits the two audiences, and states why this is economical on Stellar
specifically. It is a server component, so it is static and indexable.

Also fixes page metadata, which was still the create-next-app default:
"Create Next App" / "Generated by create next app" — the title every visitor
to the live dashboard has seen so far. Replaced with a real title,
description, and OpenGraph/Twitter card data.

README links updated: / is no longer the dashboard.

41 tests pass; lint, tsc, and build clean. Verified in a browser: landing
renders, nav to /dashboard works and still shows the real indexed payments,
zero console errors.
