+++ apps/web/src/app/layout.tsx
@@ -12,9 +12,16 @@ const geistMono = Geist_Mono({
   subsets: ["latin"],
 });
 
+const title = "Accensa — provable payments for x402 sellers on Stellar";
+const description =
+  "Agents prove they were charged correctly. Merchants refund without becoming custodians. Receipts anchored on Stellar, verifiable by anyone without an account.";
+
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title,
+  description,
+  metadataBase: new URL("https://accensa-dashboard.vercel.app"),
+  openGraph: { title, description, type: "website", siteName: "Accensa" },
+  twitter: { card: "summary_large_image", title, description },
 };
 
 export default function RootLayout({
